requirements.txt
streamlit>=1.35.0

langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0
langchain-classic>=0.3.0
langchain-chroma>=0.2.0
langchain-huggingface>=0.1.0
langchain-ollama>=0.2.0

chromadb>=0.5.0
pypdf>=4.0.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2

azure-ai-documentintelligence>=1.0.0

langchain-openai>=0.2.0

python-dotenv>=1.0.0
